+++ src/qemu/qemu_monitor_text.h
@@ -79,7 +79,7 @@ int qemuMonitorTextExpirePassword(qemuMonitorPtr mon,
                                   const char *protocol,
                                   const char *expire_time);
 int qemuMonitorTextSetBalloon(qemuMonitorPtr mon,
-                              unsigned long newmem);
+                              unsigned long long newmem);
 int qemuMonitorTextSetCPU(qemuMonitorPtr mon, int cpu, bool online);
 
 int qemuMonitorTextEjectMedia(qemuMonitorPtr mon,
